Extract byline into separate component
