+++ packages/frontend/amp/components/TopMeta.tsx
@@ -9,22 +9,8 @@ import TwitterIcon from '@guardian/pasteup/icons/twitter.svg';
 import { ShareIcons } from '@frontend/amp/components/ShareIcons';
 import { ArticleModel } from '@frontend/amp/pages/Article';
 import { MainMedia } from '@frontend/amp/components/MainMedia';
-import { bylineTokens } from '@frontend/amp/lib/byline-tokens';
 import Star from '@guardian/pasteup/icons/star.svg';
-
-const bylineStyle = (pillar: Pillar) => css`
-    ${headline(2)};
-    color: ${pillarPalette[pillar].main};
-    padding-bottom: 8px;
-    font-style: italic;
-
-    a {
-        font-weight: 700;
-        color: ${pillarPalette[pillar].main};
-        text-decoration: none;
-        font-style: normal;
-    }
-`;
+import { Byline } from '@frontend/amp/components/Byline';
 
 const meta = css`
     @supports (display: grid) {
@@ -204,32 +190,6 @@ const Headline: React.FC<{
             />
         </div>
     );
-};
-
-const Byline: React.FC<{
-    byline: string;
-    tags: TagType[];
-    pillar: Pillar;
-    guardianBaseURL: string;
-}> = ({ byline, tags, pillar, guardianBaseURL }) => {
-    const contributorTags = tags.filter(tag => tag.type === 'Contributor');
-    const tokens = bylineTokens(byline, contributorTags);
-
-    const linkedByline = tokens.map(token => {
-        const matchedTag = contributorTags.find(tag => tag.title === token);
-
-        if (matchedTag) {
-            return (
-                <a href={`${guardianBaseURL}/${matchedTag.id}`}>
-                    {matchedTag.title}
-                </a>
-            );
-        }
-
-        return token;
-    });
-
-    return <div className={bylineStyle(pillar)}>{linkedByline}</div>;
 };
 
 export const TopMeta: React.FC<{
